move channelmessage to chain.rs and rename
